Feature 13.1 soap messages

+++ car-service/src/main/java/xws/team16/carservice/dto/CarDTO.java
@@ -17,7 +17,7 @@ public class CarDTO {
     private Long carClassId;
     private Long gearboxId;
     private Long userId;
-    private int kilometrage;
+    private double kilometrage;
     private int numberOfChildSeats;
     private boolean hasAndroid;
     private List<String> images;
